+++ pkg/server/k8s/client.go
@@ -623,12 +623,12 @@ func (k *Client) HasIngress(namespace, appName string) (bool, error) {
 	return true, nil
 }
 
-func (k *Client) createIngress(namespace, appName string, vHosts []string, reserveStaticIp bool, ingressClass string) error {
+func (k *Client) createIngress(namespace, appName string, vHosts []string, ingressClass string) error {
 	kc, err := k.buildClient()
 	if err != nil {
 		return err
 	}
-	igsSpec := ingressSpec(namespace, appName, vHosts, reserveStaticIp)
+	igsSpec := ingressSpec(namespace, appName, vHosts)
 	_, err = kc.ExtensionsV1beta1().Ingresses(namespace).Create(igsSpec)
 	if err != nil {
 		return errors.Wrap(err, "create ingress failed")
@@ -644,7 +644,7 @@ func (k *Client) createIngress(namespace, appName string, vHosts []string, reser
 	return nil
 }
 
-func (k *Client) UpdateIngress(namespace, name string, vHosts []string, reserveStaticIp bool) error {
+func (k *Client) UpdateIngress(namespace, name string, vHosts []string) error {
 	kc, err := k.buildClient()
 	if err != nil {
 		return err
@@ -655,7 +655,7 @@ func (k *Client) UpdateIngress(namespace, name string, vHosts []string, reserveS
 	if err != nil {
 		return errors.Wrap(err, "get ingress failed")
 	}
-	newSpec := ingressSpec(namespace, name, vHosts, reserveStaticIp)
+	newSpec := ingressSpec(namespace, name, vHosts)
 	old.Spec.Rules = newSpec.Spec.Rules
 	_, err = kc.ExtensionsV1beta1().Ingresses(namespace).Update(old)
 	return errors.Wrap(err, "update ingress failed")
@@ -699,7 +699,7 @@ func (c *Client) IngressAnnotations(namespace, ingName string) (map[string]strin
 }
 
 // ExposeDeploy creates a service and/or a ingress if needed
-func (k *Client) ExposeDeploy(namespace, appName, svcType, portName string, vHosts []string, reserveStaticIp bool, ingressClass string, w io.Writer) error {
+func (k *Client) ExposeDeploy(namespace, appName, svcType, portName string, vHosts []string, ingressClass string, w io.Writer) error {
 	hasSrv, err := k.hasService(namespace, appName)
 	if err != nil {
 		return err
@@ -721,7 +721,7 @@ func (k *Client) ExposeDeploy(namespace, appName, svcType, portName string, vHos
 	}
 	if !hasIgs {
 		fmt.Fprintln(w, "Creating ingress")
-		if err := k.createIngress(namespace, appName, vHosts, reserveStaticIp, ingressClass); err != nil {
+		if err := k.createIngress(namespace, appName, vHosts, ingressClass); err != nil {
 			return err
 		}
 	}